abstract out duplicateNotebook
